% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrier.R
\name{plotBrier}
\alias{plotBrier}
\title{Time-dependent Brier scores}
\usage{
plotBrier(
  object,
  survObj.new = NULL,
  method = "mean",
  times = NULL,
  subgroup = 1
)
}
\arguments{
\item{object}{fitted object obtained with \code{BayesSurvive}}

\item{survObj.new}{a list containing observed data from new subjects with
components \code{t}, \code{di}, \code{X}}

\item{method}{option to use the posterior mean (\code{"mean"}) of coefficients
for prediction or Bayesian model averaging (\code{"BMA"}) for prediction}

\item{times}{maximum time point to evaluate the prediction}

\item{subgroup}{index of the subgroup in \code{survObj.new} for prediction.
Default value is 1}
}
\description{
Predict time-dependent Brier scores based on Cox regression models
}
\examples{

library("BayesSurvive")
set.seed(123)

# Load the example dataset
data("simData", package = "BayesSurvive")

dataset <- list(
  "X" = simData[[1]]$X,
  "t" = simData[[1]]$time,
  "di" = simData[[1]]$status
)

# Initial value: null model without covariates
initial <- list("gamma.ini" = rep(0, ncol(dataset$X)))
# Hyperparameters
hyperparPooled <- list(
  "c0"     = 2, # prior of baseline hazard
  "tau"    = 0.0375, # sd for coefficient prior
  "cb"     = 20, # sd for coefficient prior
  "pi.ga"  = 0.02, # prior variable selection probability for standard Cox models
  "a"      = -4, # hyperparameter in MRF prior
  "b"      = 0.1, # hyperparameter in MRF prior
  "G"      = simData$G # hyperparameter in MRF prior
)

\donttest{
# run Bayesian Cox with graph-structured priors
fit <- BayesSurvive(
  survObj = dataset, hyperpar = hyperparPooled,
  initial = initial, nIter = 100
)
# predict survival probabilities of the train data
library("GGally")
plotBrier(fit, survObj.new = dataset)
}

}
\keyword{survival}
