% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData}
\alias{simData}
\title{Simulated survival data}
\format{
An object of class \code{list} of length 3.
}
\usage{
simData
}
\description{
Simulated data set for a quick test. The data set is a list with six
components:  covariates \code{"X"}, survival times \code{"time"},
event status \code{"status"}. The R code for generating the simulated data
is given in the Examples.
}
\concept{datasets}
\keyword{datasets}
