% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePara.R
\name{UpdateGamma}
\alias{UpdateGamma}
\title{Subfunctions to update parameters}
\usage{
UpdateGamma(sobj, hyperpar, ini, S, method, MRF_G, MRF_2b)
}
\arguments{
\item{sobj}{a list containing observed data}

\item{hyperpar}{a list containing prior parameter values}

\item{ini}{a list containing prior parameters' initial values}

\item{S}{the number of subgroups}

\item{method}{a method option from
\code{c("Pooled", "CoxBVSSL", "Sub-struct", "Subgroup")}}

\item{MRF_G}{logical value. \code{MRF_G = TRUE} is to fix the MRF graph which
is provided in the argument \code{hyperpar}, and \code{MRF_G = FALSE} is to
use graphical model for leanring the MRF graph}

\item{MRF_2b}{two different b in MRF prior for subgraphs G_ss and G_rs}
}
\value{
An object of ...
}
\description{
This contains subfunctions to update parameters gammas, betas, baseline
hazard and graph learning parameters
}
\details{
BayesSurvive
}
\examples{

# Load the example dataset

}
